Enable C2H status writes
The driver expects now for every IRQ a status write to the completion
ring buffer. This way we do not "lose" IRQs if handling takes to long.
Depending on the number of status writes since the last IRQ the ISR
updates the eventfd accordingly.

+++ runtime/kernel/pcie/pcie_device.h
@@ -88,7 +88,7 @@ struct tlkm_pcie_device {
 #if LINUX_VERSION_CODE < KERNEL_VERSION(4, 8, 0)
 	struct msix_entry *msix_entries;
 #endif
-	uint32_t *cmpt_ring;
+	uint64_t *cmpt_ring;
 	dma_addr_t cmpt_ring_addr;
 };
